extract consumererror into per-signal packages
