+++ consumer/conslog/conslogerror/signalerrors_test.go
@@ -1,0 +1,43 @@
+package conslogerror
+
+import (
+	"errors"
+	"testing"
+
+	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/require"
+
+	"go.opentelemetry.io/collector/pdata/testdata"
+)
+
+func TestLogs(t *testing.T) {
+	td := testdata.GenerateLogs(1)
+	err := errors.New("some error")
+	logsErr := NewLogs(err, td)
+	assert.Equal(t, err.Error(), logsErr.Error())
+	var target Logs
+	assert.False(t, errors.As(nil, &target))
+	assert.False(t, errors.As(err, &target))
+	assert.True(t, errors.As(logsErr, &target))
+	assert.Equal(t, td, target.Data())
+}
+
+func TestLogs_Unwrap(t *testing.T) {
+	td := testdata.GenerateLogs(1)
+	var err error = testErrorType{"some error"}
+	// Wrapping err with error Logs.
+	logsErr := NewLogs(err, td)
+	target := testErrorType{}
+	require.NotEqual(t, err, target)
+	// Unwrapping logsErr for err and assigning to target.
+	require.True(t, errors.As(logsErr, &target))
+	require.Equal(t, err, target)
+}
+
+type testErrorType struct {
+	s string
+}
+
+func (t testErrorType) Error() string {
+	return ""
+}
